+++ include/smc_squared.hpp
@@ -42,7 +42,7 @@ private:
     vector<VectorXd> theta_x_prop,theta_x;
     vector<Rect> estimates;
     int n_particles,m_particles,fixed_lag,mcmc_steps;
-    vector<double>  theta_weights;
+    vector<float>  theta_weights;
     Haar haar;
     bool initialized;
